Added settings.json example info and expanded platform support
Added a line of documentation about a `settings.json` example to both the `ROBdk97.XmlDocToMd` namespace in `AssemblyDoc.cs` and the project's `README.md`. This example can be used to filter out files and namespaces. Also, a new `RuntimeIdentifiers` element was added to the `ROBdk97.XmlDocToMd.csproj` file, expanding the project's compatibility to `win-x64`, `linux-x64`, and `linux-arm64` platforms.