Added new XMLtoMD configuration and updated build commands
Added a new configuration named `XMLtoMD` to the project file `ROBdk97.XmlDocToMd.csproj` for building the project in a specific way. Modified the `PostBuild` command in the same project file to execute a certain command when the `ConfigurationName` is `XMLtoMD`. Updated the solution file `ROBdk97.XmlDocToMd.sln` to include the new `XMLtoMD` configuration for both `Any CPU` and `x64` platforms.