Refactor PostBuild target in ROBdk97.XmlDocToMd.csproj
The `PostBuild` target in the `ROBdk97.XmlDocToMd.csproj` file has been refactored. The condition to execute the `Exec` command, which was previously checked within the command itself, is now checked before the command is executed using the `Condition` attribute of the `Target` element. This change improves code readability and maintainability.