Updated post-build event command in `AssemblyDoc.cs` and `README.md`
Updated the example of a Visual Studio post-build event command in both `AssemblyDoc.cs` and `README.md` files. The previous command was written in batch script and was designed to execute when the configuration name was "Release". The updated command is written in XML and is designed to execute when the configuration name is 'XMLtoMD'. The new command is wrapped in a CDATA section in the `AssemblyDoc.cs` file to preserve the XML syntax.