Added documentation note in AssemblyDoc.cs
Added a line of documentation in the AssemblyDoc.cs file under the namespace ROBdk97.XmlDocToMd. The new line provides a notice about the tool's use of Reflection to add extra information to the XML documentation, specifying that the DLLs of the project must be located in the same directory as the XML documentation files.