Updated text emphasis in documentation
Updated the emphasis of a notice in the XML documentation. The text "Notice: As this tool uses Reflection to add additional information to the XML documentation, the DLL´s of the project must be in the same directory as the XML documentation files." was updated to "**Notice:** As this tool uses Reflection to add additional information to the XML documentation, the DLL´s of the project must be in the same directory as the XML documentation files." This change was made to emphasize the notice by making it bold.